# API_Gateway_Authentication_Service/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subscription Management API",
        "version": "1.0.0-SNAPSHOT",
...
        "contact": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            }
        },
        "license": {
            "name": "MIT",
            "url": "https://spdx.org/licenses/MIT.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "notFoundResponseText": {
                "type": "string",
                "maxLength": 2000,
                "pattern": "^[ \\w\\.!?,;:\\-=><\\/&%\\$\\+\\*~#]*$",
                "example": "Not Found"
            },
...
            "badRequestText": {
                "type": "string",
                "maxLength": 2000,
                "pattern": "^[ \\w\\.!?,;:\\-=><\\/&%\\$\\+\\*~#]*$",
                "example": "Bad Request"
            },
            "forbiddenText": {
                "type": "string",
                "maxLength": 2000,
                "pattern": "^[ \\w\\.!?,;:\\-=><\\/&%\\$\\+\\*~#]*$",
                "example": "Forbidden"
            },
            "unauthorizedText": {
                "type": "string",
                "maxLength": 2000,
                "pattern": "^[ \\w\\.!?,;:\\-=><\\/&%\\$\\+\\*~#]*$",
                "example": "Unauthorized"
            },
...
            "attrPlanID": {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z]+[\\-\\.\\w]*$",
                "example": "Free"
            },
...
            "planData": {
                "description": "a plan profile",
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "planID",
                    "monthlyFee",
                    "annualFee",
                    "numberOfMinutes",
                    "numberOfDevice",
                    "musicCollections",
                    "musicSuggestions"
                ],
                "properties": {
                    "planID": {
                        "$ref": "#/components/schemas/attrPlanID"
                    },
                    "monthlyFee": {
                        "$ref": "#/components/schemas/attrMonthlyFee"
                    },
                    "annualFee": {
                        "$ref": "#/components/schemas/attrAnnualFee"
                    },
                    "numberOfMinutes": {
                        "$ref": "#/components/schemas/attrNumberOfMinutes"
                    },
                    "numberOfDevice": {
                        "$ref": "#/components/schemas/atrNumberOfDevice"
                    },
                    "musicCollections": {
                        "$ref": "#/components/schemas/attrMusicCollections"
                    },
                    "musicSuggestions": {
                        "$ref": "#/components/schemas/attributeMusicSuggestions"
                    },
                    "annualFeeChange": {
                        "$ref": "#/components/schemas/attrAnnualFeeChange"
                    },
                    "monthlyFeeChange": {
                        "$ref": "#/components/schemas/attrMonthlyFeeChange"
                    },
                    "status": {
                        "$ref": "#/components/schemas/attributeStatus"
                    }
                }
            },
...
            "subscriptionID": {
                "description": "The id of the subscription",
                "type": "string",
                "maxLength": 8,
                "pattern": "^[- \\w\\.!?,;:-=><\\/()&%$+*~#]*$",
                "example": "hbad0z1y"
            },
...
            "plansDataCollection": {
                "description": "collection of plans",
                "type": "array",
                "maxItems": 500,
                "items": {
                    "$ref": "#/components/schemas/planData"
                }
            },
...
            "planItem": {
                "description": "a plan",
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/planData"
                    },
                    "links": {
                        "$ref": "#/components/schemas/links"
                    }
                }
            },
...
            "upgradeSubscriptionPlanInput": {
                "description": "input for upgrading a subscription plan",
                "type": "object",
                "additionalProperties": false,
                "required": ["planId"],
                "properties": {
                    "planId": {
                        "$ref": "#/components/schemas/attrPlanID"
                    }
                }
            },
...
            "changePlanDetailsInput": {
                "description": "input for changing plan details",
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "monthlyFee": {
                        "$ref": "#/components/schemas/attrMonthlyFee"
                    },
                    "numberOFDevices": {
                        "$ref": "#/components/schemas/atrNumberOfDevice"
                    },
                    "devices": {
                        "$ref": "#/components/schemas/atrNumberOfDevice"
                    },
                    "promote": {
                        "$ref": "#/components/schemas/atrPromote"
                    }
                }
            },
            "newPlanInput": {
                "description": "input for creating a new plan",
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "planID",
                    "monthlyFee",
                    "annualFee",
                    "numberOfMinutes",
                    "numberOfDevice",
                    "musicCollections",
                    "musicSuggestions"
                ],
                "properties": {
                    "planID": {
                        "$ref": "#/components/schemas/attrPlanID"
                    },
                    "monthlyFee": {
                        "$ref": "#/components/schemas/attrMonthlyFee"
                    },
                    "annualFee": {
                        "$ref": "#/components/schemas/attrAnnualFee"
                    },
                    "numberOfMinutes": {
                        "$ref": "#/components/schemas/attrNumberOfMinutes"
                    },
                    "numberOfDevice": {
                        "$ref": "#/components/schemas/atrNumberOfDevice"
                    },
                    "musicCollections": {
                        "$ref": "#/components/schemas/attrMusicCollections"
                    },
                    "musicSuggestions": {
                        "$ref": "#/components/schemas/attributeMusicSuggestions"
                    }
                }
            },
...
        },
        "responses": {
...
            "OK": {
                "description": "Ok.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/serverResponseText"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Not Found.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/notFoundResponseText"
                        }
                    }
                }
            },
...
            "BadRequest": {
                "description": "Bad Request.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/badRequestText"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized. You must first authenticate in order to perform this request.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/unauthorizedText"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden. You do not have enough permissions to perform this request.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/forbiddenText"
                        }
                    }
                }
            },
...
            "OkPlanCollection": {
                "description": "the plan collection",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/plansDataCollection"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/plansCollection"
                        }
                    }
                }
            },
            "OkPlanItem": {
                "description": "the plan item",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/planData"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/planItem"
                        }
                    }
                },
                "links": {
                    "self": {
                        "operationId": "getPlan",
                        "parameters": {
                            "id": "$response.body#/id"
                        }
                    },
                    "collection": {
                        "operationId": "getPlans"
                    },
                    "update": {
                        "operationId": "updatePlan",
                        "parameters": {
                            "id": "$response.body#/id"
                        }
                    }
                }
            },
            "PreconditionFailedPlanItem": {
                "description": "if-match header caused conflict",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/planData"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/planItem"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
    "paths": {
...
        "/subscription/{subscriptionID}/plan": {
            "get": {
                "description": "returns the plan of one subscription",
                "summary": "returns the plan of one subscription",
                "tags": ["Plans"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "subscriptionID",
                        "schema": {
                            "$ref": "#/components/schemas/subscriptionID"
                        },
                        "description": "the subscription ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanItem"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                }
            },
            "patch": {
                "description": "updates the plan of one subscription",
                "summary": "updates the plan of one subscription",
                "tags": ["Subscriptions"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "subscriptionID",
                        "schema": {
                            "$ref": "#/components/schemas/subscriptionID"
                        },
                        "description": "the subscription ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "the plan data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/upgradeSubscriptionPlanInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanItem"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "412": {
                        "$ref": "#/components/responses/PreconditionFailedPlanItem"
                    }
                }
            }
        },
...
        "/plans": {
            "get": {
                "description": "returns all plans",
                "summary": "returns all plans",
                "tags": ["Plans"],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "first",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "maximum": 500,
                            "format": "int32"
                        },
                        "description": "the first user to return",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "n",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 6,
                            "maximum": 500,
                            "format": "int32"
                        },
                        "description": "the number of users to return",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanCollection"
                    }
                }
            }
        },
        "/plans/{planID}": {
            "get": {
                "description": "returns a plan",
                "summary": "returns a plan",
                "tags": ["Plans"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "planID",
                        "schema": {
                            "$ref": "#/components/schemas/attrPlanID"
                        },
                        "description": "the plan ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanItem"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "description": "creates or updates a plan",
                "summary": "creates or updates a plan",
                "tags": ["Plans"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "planID",
                        "schema": {
                            "$ref": "#/components/schemas/attrPlanID"
                        },
                        "description": "the plan ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "the plan data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/newPlanInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanItem"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "412": {
                        "$ref": "#/components/responses/PreconditionFailedPlanItem"
                    }
                }
            },
            "delete": {
                "description": "deletes a plan",
                "summary": "deletes a plan",
                "tags": ["Plans"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "planID",
                        "schema": {
                            "$ref": "#/components/schemas/attrPlanID"
                        },
                        "description": "the plan ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OK"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "412": {
                        "$ref": "#/components/responses/PreconditionFailedPlanItem"
                    }
                }
            },
            "patch": {
                "description": "updates a plan",
                "summary": "updates a plan",
                "tags": ["Plans"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "planID",
                        "schema": {
                            "$ref": "#/components/schemas/attrPlanID"
                        },
                        "description": "the plan ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "the plan data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/changePlanDetailsInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanItem"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "412": {
                        "$ref": "#/components/responses/PreconditionFailedPlanItem"
                    }
                }
            }
        },
...
    },
...
}